Update for October 12
Some things are still a work in progress. I tried to play around with a rotated header, but could not move it into the position I wanted. It also clips above my menu with a fixed position, even though it is located well below the header in my code. Any help would be appreciated!